Stop catching parser errors
Having any exception on the parser stack trapped requires any
callback/event going back to the user to be wrapped in process.nextTick
in order for it to not be trapped by try...catch block which is very
annoying.

This also makes it hard to guarantee row streaming pause() / resume() to
work properly as invoking the 'result' events in a nextTick means that
other 'result' events may fire, even if the first one calls pause()
(because the other ones are already scheduled as nextTicks).

Even so it would be nice to allow the user to handle potentially unknown
problems in the parser gracefully, it is ultimatively not worth the
tradeoffs that are involved. Going forward any unknown parser defects
will cause hard exceptions which may only be trapped by using domains or
'uncaughtException'.

This patch is also done in preparation for implementing exception
safety, so that it becomes possible to use a shared mysql connection
across multiple domains.