Splitting the changeUser test into separate user change and db change
In b3220d3, the change-user-fatal-error test was modified in order to run
on Travis as well. However, according to the comment in the code, changing
to an invalid user doesn't seem to cause an error on Travis side, so this
test was modified to change the database instead. This only seems to make
limited sense to me (this obviously is not a limitation of the code, but
of Travis), so I basically undid that commit for this test.
However, I added another test that actually changes the database, which
also runs on Travis.
In addition, I disabled the check for a particular error code, since they
seem to vary according to database configuration and setup, just the `fatal`
property is checked.