Optimize SqlString.format
I was looking for ways to speed up mysql.format after noticing that the performance was very slow when attempting to format a query with 100,000+ values. I came up with a couple of improvements, but by far the most significant is using an iterator to step through the values array instead of Array.shift(). See: http://jsperf.com/some-array-reading-comparisons/6