Update documentation around custom formatters.
The example worked for me but since 99.9% of the existing project was using ? placeholders I tidied the Function example up to first check if the SQL statement used ? and then customised for that before checking if it should be named parameters.  Also we have one SQL that seems to be using := assignment, I have added a condition for it to not effect that.